refactor(endpoints): move stack inspect to edge handler
